% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaters-package.R
\docType{package}
\name{skaters-package}
\alias{skaters}
\alias{skaters-package}
\title{skaters: Online Distributional Time-Series Forecasting by Conjugation}
\description{
Port of the Python 'skaters' package: an online, univariate, distributional forecaster built from invertible transforms over a Gaussian-mixture leaf. Every implemented component is verified against the reference implementation's parity vectors to 1e-6.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/microprediction/skaters-r}
  \item \url{https://skaters.microprediction.org}
  \item Report bugs at \url{https://github.com/microprediction/skaters-r/issues}
}

}
\author{
\strong{Maintainer}: Peter Cotton \email{peter.cotton@microprediction.com} (\href{https://orcid.org/0000-0003-1832-2924}{ORCID})

Authors:
\itemize{
  \item Peter Cotton \email{peter.cotton@microprediction.com} (\href{https://orcid.org/0000-0003-1832-2924}{ORCID})
}

}
\keyword{internal}
